Extracted Time to Core module, fixed tooling and types.
