+++ ts/Core/Options.d.ts
@@ -19,6 +19,7 @@ import type ColorString from './Color/ColorString';
 import type CSSObject from './Renderer/CSSObject';
 import type { SeriesTypePlotOptions } from './Series/SeriesType';
 import type { SymbolKey } from './Renderer/SVG/SymbolType';
+import type { LangOptionsCore } from '../Shared/LangOptionsCore';
 
 /* *
  *
@@ -40,18 +41,6 @@ export interface LangOptions extends LangOptionsCore {
     zoomIn?: string;
     zoomOut?: string;
 }
-
-export interface LangOptionsCore {
-    decimalPoint?: string;
-    invalidDate?: string;
-    locale?: string|Array<string>;
-    months?: Array<string>;
-    shortMonths?: Array<string>;
-    shortWeekdays?: Array<string>;
-    thousandsSep?: string;
-    weekdays?: Array<string>;
-}
-
 export interface LoadingOptions {
     hideDuration?: number;
     labelStyle?: CSSObject;
