+++ ts/DataGrid/Options.ts
@@ -25,7 +25,7 @@ import type DataTable from '../Data/DataTable';
 import type DataTableOptions from '../Data/DataTableOptions';
 import type Cell from './Table/Cell';
 import type Column from './Table/Column';
-import type { LangOptionsCore } from '../Core/Options';
+import type { LangOptionsCore } from '../Shared/LangOptionsCore';
 
 
 /* *
