+++ ts/Shared/LangOptionsCore.d.ts
@@ -1,0 +1,26 @@
+/* *
+ *
+ *  (c) 2010-2024 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface LangOptionsCore {
+    decimalPoint?: string;
+    invalidDate?: string;
+    locale?: string|Array<string>;
+    months?: Array<string>;
+    shortMonths?: Array<string>;
+    shortWeekdays?: Array<string>;
+    thousandsSep?: string;
+    weekdays?: Array<string>;
+}
